src/requirements.txt
asgiref==3.11.0
distlib==0.4.0
Django==6.0
filelock==3.20.1
platformdirs==4.5.1
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.4
virtualenv==20.35.4
